refactor: split standalone e2e scenarios
