+++ tests/e2e_standalone/test_mcp_protocol.py
@@ -1,0 +1,393 @@
+"""Maya standalone E2E tests."""
+
+from __future__ import annotations
+
+import pytest
+
+from ._support import _mcp_list_all_tools, _mcp_post, _new_scene
+
+pytestmark = pytest.mark.e2e
+
+
+class TestMcpHttpConnectivity:
+    """MCP server responds correctly to real HTTP JSON-RPC requests.
+
+    Uses a class-scoped server fixture so all tests share one port.
+    """
+
+    @pytest.fixture(autouse=True, scope="class")
+    def _start_server(self, request):
+        from dcc_mcp_maya.server import MayaMcpServer
+
+        _new_scene()
+        server = MayaMcpServer(port=0)
+        server.register_builtin_actions()
+        handle = server.start()
+        request.cls._mcp_url = handle.mcp_url()
+        yield
+        server.stop()
+
+    def test_initialize_handshake(self):
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 1,
+                "method": "initialize",
+                "params": {
+                    "protocolVersion": "2025-03-26",
+                    "capabilities": {},
+                    "clientInfo": {"name": "e2e-test", "version": "1.0"},
+                },
+            },
+        )
+        assert code == 200
+        result = body["result"]
+        assert result["protocolVersion"] == "2025-03-26"
+        assert result["serverInfo"]["name"] == "maya-mcp"
+
+    def test_tools_list_shows_core_and_stubs(self):
+        """tools/list returns core discovery tools + skill stubs (progressive loading).
+
+        In the progressive (lazy) loading model, tools/list returns three layers:
+        1. Core discovery tools (search_skills, list_skills, get_skill_info, load_skill, ...)
+        2. Already-loaded skill tools with full input_schema
+        3. Unloaded skill stubs as ``__skill__<name>`` with minimal description
+
+        Core paginates the response (~32 tools/page); we aggregate every
+        page before asserting so a future tool registration that shifts
+        alphabetic ordering cannot silently push stubs off page 1.
+        """
+        tools = _mcp_list_all_tools(self._mcp_url, request_id=2)
+        names = {t["name"] for t in tools}
+
+        # Layer 1: core discovery tools must always be present
+        for core_tool in ("search_skills", "list_skills", "get_skill_info", "load_skill", "unload_skill"):
+            assert core_tool in names, f"Core tool {core_tool!r} missing from tools/list"
+
+        # Layer 3: unloaded skills appear as stubs (__skill__<name>)
+        stub_names = {n for n in names if n.startswith("__skill__")}
+        assert len(stub_names) >= 1, "Expected at least one skill stub in progressive mode"
+
+        # Stubs should have minimal schema (no or empty input_schema)
+        for t in tools:
+            if t["name"].startswith("__skill__"):
+                schema = t.get("inputSchema", {})
+                # Stubs have empty or minimal input_schema
+                assert schema.get("properties", {}) == {} or "properties" not in schema, (
+                    f"Stub {t['name']} should not have full input_schema"
+                )
+
+    def test_load_skill_replaces_skill_stub(self):
+        """Calling load_skill removes a __skill__ stub and exposes its tools.
+
+        In progressive (minimal) mode, undiscovered skills appear as
+        ``__skill__<name>`` stubs.  Loading a skill replaces the stub with
+        the skill's real tools (from groups that have ``default_active: true``)
+        and any ``__group__`` stubs for inactive groups.
+        """
+        # Grab baseline tool names (paginated aggregate — see docstring on _mcp_list_all_tools).
+        before_names = {t["name"] for t in _mcp_list_all_tools(self._mcp_url, request_id=10)}
+
+        # Pick any __skill__ stub to load
+        skill_stubs = sorted(n for n in before_names if n.startswith("__skill__"))
+        assert len(skill_stubs) >= 1, "Need at least one __skill__ stub to test load_skill"
+
+        skill_stub = skill_stubs[0]
+        skill_name = skill_stub.replace("__skill__", "")
+
+        # Load the skill via tools/call
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 11,
+                "method": "tools/call",
+                "params": {"name": "load_skill", "arguments": {"skill_name": skill_name}},
+            },
+        )
+        assert code == 200
+
+        # tools/list (aggregated) should no longer contain the __skill__ stub
+        after_names = {t["name"] for t in _mcp_list_all_tools(self._mcp_url, request_id=12)}
+        assert skill_stub not in after_names, f"Stub {skill_stub} should be removed after load_skill"
+
+    def test_activate_tool_group_replaces_group_stub(self):
+        """Calling activate_tool_group removes a __group__ stub and exposes its tools.
+
+        In progressive (minimal) mode, loaded skills may have groups marked
+        ``default_active: false`` that appear as ``__group__<name>`` stubs.
+        Activating such a group replaces the stub with the group's real tools.
+        """
+        # Grab baseline tool names (paginated aggregate)
+        before_names = {t["name"] for t in _mcp_list_all_tools(self._mcp_url, request_id=20)}
+
+        # Pick any __group__ stub to activate
+        group_stubs = sorted(n for n in before_names if n.startswith("__group__"))
+        assert len(group_stubs) >= 1, "Need at least one __group__ stub to test activate_tool_group"
+
+        group_stub = group_stubs[0]
+        group_name = group_stub.replace("__group__", "")
+
+        # Activate the tool group via tools/call
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 21,
+                "method": "tools/call",
+                "params": {
+                    "name": "activate_tool_group",
+                    "arguments": {"group": group_name},
+                },
+            },
+        )
+        assert code == 200
+
+        # tools/list (aggregated) should no longer contain the __group__ stub
+        after_names = {t["name"] for t in _mcp_list_all_tools(self._mcp_url, request_id=22)}
+        assert group_stub not in after_names, f"Stub {group_stub} should be removed after activate_tool_group"
+
+        # The activated group's real tools should now be present.
+        new_tools = after_names - before_names - {n for n in after_names if n.startswith("__")}
+        assert len(new_tools) >= 1, (
+            f"Expected at least one new tool after activating group {group_name!r}, "
+            f"but no bare or prefixed tools appeared. "
+            f"before={sorted(before_names)[:5]} after={sorted(after_names)[:5]}"
+        )
+
+    @pytest.mark.xfail(
+        reason="dcc-mcp-core subprocess executor uses relative script path; known issue",
+        strict=False,
+    )
+    def test_tools_call_get_session_info_via_http(self):
+        """tools/call returns Maya session info over HTTP."""
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 3,
+                "method": "tools/call",
+                "params": {"name": "maya-scene.get_session_info", "arguments": {}},
+            },
+        )
+        assert code == 200
+        content = body["result"]["content"]
+        assert len(content) > 0
+        # Content is a list of {type, text} blocks
+        text = content[0].get("text", "")
+        assert "maya_version" in text or "success" in text
+
+    @pytest.mark.xfail(
+        reason="dcc-mcp-core subprocess executor uses relative script path; known issue",
+        strict=False,
+    )
+    def test_tools_call_create_sphere_via_http(self):
+        """tools/call create_sphere returns success in JSON response."""
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 4,
+                "method": "tools/call",
+                "params": {
+                    "name": "maya-primitives.create_sphere",
+                    "arguments": {"radius": 1.5, "name": "httpSphere"},
+                },
+            },
+        )
+        assert code == 200
+        # Verify JSON response indicates success (MCP server ran the tool)
+        content = body["result"]["content"]
+        assert len(content) > 0
+        text = content[0].get("text", "")
+        assert "success" in text or "httpSphere" in text
+
+    @pytest.mark.xfail(
+        reason="dcc-mcp-core subprocess executor uses relative script path; known issue",
+        strict=False,
+    )
+    def test_tools_call_execute_python_via_http(self):
+        """tools/call execute_python returns success in JSON response."""
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 5,
+                "method": "tools/call",
+                "params": {
+                    "name": "maya-scripting.execute_python",
+                    "arguments": {"code": "import maya.cmds as cmds; cmds.polyCube(n='httpCube')"},
+                },
+            },
+        )
+        assert code == 200
+        content = body["result"]["content"]
+        assert len(content) > 0
+        # Tool executed — response contains result text
+        text = content[0].get("text", "")
+        assert isinstance(text, str) and len(text) > 0
+
+
+class TestScriptingHttpE2E:
+    """Real MCP client ↔ in-Maya server round-trips for ``execute_python``.
+
+    These tests start an MCP server inside ``mayapy`` (the surrounding
+    Maya standalone process) and drive it via real HTTP JSON-RPC to the
+    ``/mcp`` endpoint — the same wire shape an external Codebuddy /
+    Claude / Cursor client would use.  This is stronger than the
+    in-process skill-script tests in ``tests/e2e/test_scripting_e2e.py``
+    because it exercises the full dispatch chain: transport → core
+    registry → in-process executor → Maya UI dispatcher → skill script.
+    """
+
+    @pytest.fixture(autouse=True, scope="class")
+    def _start_server_and_load_scripting(self, request):
+        # Import built-in modules
+        import os  # noqa: PLC0415
+
+        # Import local modules
+        from dcc_mcp_maya.server import MayaMcpServer  # noqa: PLC0415
+
+        # In-process skill execution requires dcc-mcp-core to trust that
+        # ``import maya.cmds`` works under the ambient interpreter
+        # (upstream issue dcc-mcp-core#231).  We are running inside
+        # mayapy where that guarantee holds, but core has no reliable
+        # way to detect that — flip the documented test escape hatch
+        # so the skill script actually executes instead of short-
+        # circuiting with an ``EXECUTION_FAILED`` envelope.
+        ambient_env = "DCC_MCP_ALLOW_AMBIENT_PYTHON"
+        previous_env = os.environ.get(ambient_env)
+        os.environ[ambient_env] = "1"
+
+        _new_scene()
+        server = MayaMcpServer(port=0)
+        server.register_builtin_actions()
+        handle = server.start()
+        request.cls._mcp_url = handle.mcp_url()
+
+        # maya-scripting lives in the minimal mode default set, but be explicit
+        # so tests do not depend on the default-skill configuration.
+        _mcp_post(
+            request.cls._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": 900,
+                "method": "tools/call",
+                "params": {
+                    "name": "load_skill",
+                    "arguments": {"skill_name": "maya-scripting"},
+                },
+            },
+        )
+        try:
+            yield
+        finally:
+            server.stop()
+            if previous_env is None:
+                os.environ.pop(ambient_env, None)
+            else:
+                os.environ[ambient_env] = previous_env
+
+    def _resolve_execute_python_tool_name(self) -> str:
+        """Return the MCP tool name the server registered for ``execute_python``.
+
+        Depending on the skill-registration mode the tool may be exposed
+        as ``execute_python`` (bare) or ``maya_scripting__execute_python``
+        (fully qualified).  Prefer the fully-qualified form if present so
+        we do not collide with any other skill that happens to declare an
+        ``execute_python`` tool.
+        """
+        _, body = _mcp_post(
+            self._mcp_url,
+            {"jsonrpc": "2.0", "id": 901, "method": "tools/list"},
+        )
+        names = [t["name"] for t in body["result"]["tools"]]
+        qualified = [n for n in names if n.endswith("__execute_python")]
+        if qualified:
+            return qualified[0]
+        if "execute_python" in names:
+            return "execute_python"
+        raise AssertionError(f"execute_python tool not exposed; saw: {sorted(names)[:40]}")
+
+    def _call(self, tool_name: str, arguments: dict, request_id: int = 1000) -> dict:
+        """Issue a ``tools/call`` and return the parsed result body."""
+        code, body = _mcp_post(
+            self._mcp_url,
+            {
+                "jsonrpc": "2.0",
+                "id": request_id,
+                "method": "tools/call",
+                "params": {"name": tool_name, "arguments": arguments},
+            },
+        )
+        assert code == 200, f"HTTP {code} from tools/call({tool_name!r})"
+        assert "result" in body, f"no result in response: {body}"
+        return body
+
+    def test_execute_python_inline_roundtrip(self):
+        """Baseline: sync ``execute_python`` completes over real HTTP."""
+        tool = self._resolve_execute_python_tool_name()
+        body = self._call(
+            tool,
+            {"code": "import maya.cmds as cmds; cmds.polyCube(n='httpInlineCube')"},
+            request_id=1001,
+        )
+        content = body["result"]["content"]
+        assert content, "tools/call returned empty content"
+        text = content[0].get("text", "")
+        # Envelope either surfaces success=true or a structured dict string.
+        assert "success" in text.lower() or "httpInlineCube" in text
+
+    def test_execute_python_captures_cmds_warning_over_http(self):
+        """Issue #151 — ``cmds.warning(...)`` must surface in the HTTP payload.
+
+        Uses a deliberately simple single-statement snippet to stay on
+        the happy path of dcc-mcp-core's parameter marshalling — the
+        subprocess executor has trouble with compound statements in
+        some environments, and that is orthogonal to what this test is
+        verifying (the Maya-output-capture hook itself).
+        """
+        tool = self._resolve_execute_python_tool_name()
+        marker = "e2e_http_warning_marker"
+        body = self._call(
+            tool,
+            {"code": "import maya.cmds as cmds; cmds.warning({!r})".format(marker)},
+            request_id=1002,
+        )
+        content = body["result"]["content"]
+        assert content, "tools/call returned empty content"
+        text = content[0].get("text", "")
+        # If the in-process executor happens to be unavailable (older
+        # cores, missing env bypass), the envelope will carry
+        # EXECUTION_FAILED without the warning text — treat that as an
+        # infrastructure skip rather than a semantic failure so this
+        # test remains useful in the common case.
+        if "EXECUTION_FAILED" in text and marker not in text:
+            pytest.skip("in-process executor unavailable in this mayapy image: {!r}".format(text[:200]))
+        assert marker in text, "cmds.warning output must reach the HTTP client (issue #151). Got: {!r}".format(
+            text[:400]
+        )
+
+    def test_execute_python_defer_roundtrip(self):
+        """Issue #153 — ``defer=True`` returns a completed envelope over HTTP.
+
+        Uses a short snippet so the deferred-tool poll loop resolves
+        within the request timeout.  This proves the DeferredToolResult
+        path is wired end-to-end (skill → executor → core poll →
+        HTTP response) without regressions.
+        """
+        tool = self._resolve_execute_python_tool_name()
+        body = self._call(
+            tool,
+            {
+                "code": "import maya.cmds as cmds; cmds.polyCube(n='httpDeferCube')",
+                "defer": True,
+                "timeout_secs": 30,
+            },
+            request_id=1003,
+        )
+        content = body["result"]["content"]
+        assert content, "deferred tools/call returned empty content"
+        text = content[0].get("text", "")
+        assert "success" in text.lower() or "httpDeferCube" in text
